fix(shared): move marco into modernjs config
