--- .vscode/settings.json
@@ -1,36 +1,0 @@
-{
-  "editor.codeActionsOnSave": {
-    "source.organizeImports.biome": "explicit"
-  },
-  "editor.defaultFormatter": "biomejs.biome",
-  "editor.formatOnSave": true,
-  "cSpell.words": [
-    "AITEST",
-    "Aliyun",
-    "aweme",
-    "bbox",
-    "bytedance",
-    "deepseek",
-    "doubao",
-    "douyin",
-    "fkill",
-    "httpbin",
-    "iconfont",
-    "jsonrepair",
-    "modelcontextprotocol",
-    "openrouter",
-    "qwen",
-    "taobao",
-    "targetcreated",
-    "Volcengine",
-    "xpaths",
-    "Yadb"
-  ],
-  "[jsonc]": {
-    "editor.defaultFormatter": "biomejs.biome"
-  },
-  "[plaintext]": {
-    "editor.defaultFormatter": "esbenp.prettier-vscode"
-  },
-  "nxConsole.generateAiAgentRules": true
-}
